Added persistent store using the local storage to hide how to session are working.
Indeed the session must always be synhronized with the user data stored in the store-redux. Thus, for each dispatch into the store, we save the redux partial into the localStorage. We load the persited state at the begining of the App. We will use a 'userSelector' applyed to the store to retrieve the user. Therefore, we will avoid unsynchronized session or inconsistent  method to retrive it.